Support transient SMART failures
As discussed in [1] some SMART errors are transient and should not
be treated as permanent.

This commit adds support for a configurable list of ATA SMART attribute
IDs, failures of which will be treated as transient. Drive health history
is still recorded and notifications are sent, but the device itself is
not marked as failed.

Fixes #374.

[1] https://github.com/AnalogJ/scrutiny/issues/374